:recycle: jotai 적용 및 컴포넌트 분리
